LayoutTests/accessibility/presentation-role-iframe-expected.txt




This tests that setting role=presentation on an iframe/object hides the AXWebArea.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


1. Content child: AXRole: AXGroup
1. Content grand child: AXRole: AXButton
2. Content child: AXRole: AXScrollArea
2. Content grand child: AXRole: AXWebArea
3. Content child: AXRole: AXGroup
3. Content grand child: AXRole: AXButton
PASS successfullyParsed is true

TEST COMPLETE
